+++ apollo-router/src/graphql.rs
@@ -1,0 +1,91 @@
+use crate::error::{FetchError, Location};
+use crate::json_ext::Object;
+use crate::json_ext::Path;
+use crate::json_ext::Value;
+use serde::{Deserialize, Serialize};
+use thiserror::Error;
+
+pub use crate::request::Request;
+pub use crate::response::Response;
+
+/// Any GraphQL error.
+#[derive(Error, Clone, Debug, Eq, PartialEq, Serialize, Deserialize, Default)]
+#[error("{message}")]
+#[serde(rename_all = "camelCase")]
+pub struct Error {
+    /// The error message.
+    pub message: String,
+
+    /// The locations of the error from the originating request.
+    pub locations: Vec<Location>,
+
+    /// The path of the error.
+    pub path: Option<Path>,
+
+    /// The optional graphql extensions.
+    #[serde(default, skip_serializing_if = "Object::is_empty")]
+    pub extensions: Object,
+}
+
+#[buildstructor::buildstructor]
+impl Error {
+    #[builder]
+    pub fn new(
+        message: String,
+        locations: Vec<Location>,
+        path: Option<Path>,
+        extensions: Option<Object>,
+    ) -> Self {
+        Self {
+            message,
+            locations,
+            path,
+            extensions: extensions.unwrap_or_default(),
+        }
+    }
+
+    pub fn from_value(service_name: &str, value: Value) -> Result<Error, FetchError> {
+        let mut object =
+            ensure_object!(value).map_err(|error| FetchError::SubrequestMalformedResponse {
+                service: service_name.to_string(),
+                reason: error.to_string(),
+            })?;
+
+        let extensions =
+            extract_key_value_from_object!(object, "extensions", Value::Object(o) => o)
+                .map_err(|err| FetchError::SubrequestMalformedResponse {
+                    service: service_name.to_string(),
+                    reason: err.to_string(),
+                })?
+                .unwrap_or_default();
+        let message = extract_key_value_from_object!(object, "message", Value::String(s) => s)
+            .map_err(|err| FetchError::SubrequestMalformedResponse {
+                service: service_name.to_string(),
+                reason: err.to_string(),
+            })?
+            .map(|s| s.as_str().to_string())
+            .unwrap_or_default();
+        let locations = extract_key_value_from_object!(object, "locations")
+            .map(serde_json_bytes::from_value)
+            .transpose()
+            .map_err(|err| FetchError::SubrequestMalformedResponse {
+                service: service_name.to_string(),
+                reason: err.to_string(),
+            })?
+            .unwrap_or_default();
+        let path = extract_key_value_from_object!(object, "path")
+            .map(serde_json_bytes::from_value)
+            .transpose()
+            .map_err(|err| FetchError::SubrequestMalformedResponse {
+                service: service_name.to_string(),
+                reason: err.to_string(),
+            })?;
+
+        Ok(Error {
+            message,
+            locations,
+            path,
+            extensions,
+        })
+    }
+}
